# Speed up "VMs & Instances in My LDAP Group" filter in /vm_or_template/explorer