Purpose or Intent
-----------------
Speeds up "VMs & Instances in My LDAP Group" index and ajax calls in the `/vm_or_template` (Services -> Workload) section of the application.


How
---
This adds `virtual_attribute` on to the `ownership_mixin`, which is used by a few classes and makes it so that MiqExpressions will evaluate filter in SQL instead of pulling down all matching records to then interate over and filter in Ruby (causes a Massive N+1 on large datasets).  This requires a join for the `virtual_attribute` to function, so the `with_miq_group` scope does that with out pulling in any extra data, which is added in the controller which uses `MiqReport#paged_view_search` with an `MiqExpression`.


Metrics
-------
The data set for these metrics used a DB with approximately 20k `Vm` records.


**Before (Regular User)**

Started POST "/vm_or_template/tree_select/?id=:id"...
Completed 200 OK in 7242ms (Views: 2.0ms | ActiveRecord: 884.1ms)

|      ms |   bytes |   objects | queries | query (ms) |   rows |
|    ---: |    ---: |      ---: |    ---: |       ---: |   ---: |
| 6,043.1 |     N/A | 3,769,724 |      44 |    4,158.3 | 29,709 |


**After (Regular User)**

Started POST "/vm_or_template/tree_select/?id=:id"...
Completed 200 OK in 823ms (Views: 2.2ms | ActiveRecord: 46.6ms)

|      ms |   bytes |   objects | queries | query (ms) | rows |
|    ---: |    ---: |      ---: |    ---: |       ---: | ---: |
|   659.0 |     N/A |   816,802 |      30 |      131.9 |  397 |


**Before (Admin)**

Started POST "/vm_or_template/tree_select/?id=:id"...
Completed 200 OK in 94316ms (Views: 1.1ms | ActiveRecord: 13351.1ms)

|        ms | bytes |    objects | queries | query (ms) |    rows |
|      ---: |  ---: |       ---: |    ---: |       ---: |    ---: |
|  93,770.4 |   N/A | 57,852,016 |     112 |   78,603.0 | 349,084 |



**After (Admin)**

Started POST "/vm_or_template/tree_select/?id=:id"...
Completed 200 OK in 283ms (Views: 1.0ms | ActiveRecord: 12.6ms)

|      ms | bytes |   objects | queries | query (ms) | rows |
|    ---: |  ---: |      ---: |    ---: |       ---: | ---: |
|   190.2 |   N/A |    91,554 |      20 |       20.3 |   14 |


Additional Notes
----------------

This change drastically reduces the number of `rows` returns, time spend churning in ruby (sometimes to return nothing), decreases bandwidth and memory footprints, but the implementation in the controller is a bit hacky, and most likely requires some other changes now in other places to restore previous functionality.  It also attaches the `with_miq_group` scope to all other calls.  This should be fine for the most part, but could cause other regressions.  Not sure of a better way to limit the inclusion of this scope using the current mechanisms available without some other heavier refactoring required.


_Update:_

> It also attaches the `with_miq_group` scope to all other calls.

This is no longer the case as I have added controller methods to only add that scope when the MiqExpression "expression" includes the `owned_by_current_ldap_group` field.