@NickLaMuro 2016-08-05
FYI to the Reviewer:  Since the `owned_by_current_ldap_group` now requires `with_miq_groups` there are probably a lot more places where the controllers will need to be changed (or at least higher up the inheritance stack).  I skipped that for now incase code review ends up leading to a better implementation of the same SQL, and will fix before merging.