@NickLaMuro 2016-08-10
@Fryguy I apologize for the delay on getting the updates to this PR out, but the fixes you requested should be here now.

Some items of note:

* Changed the test in the first commit to be a `shared_example` instead of a mixin test, since what we care about is that the models that inherit this code behave the same.  This triples the amount of time this is run, since I include it in three model tests that include this module, but it will make sure the same scenarios are run in each and all three work. ( @chrisarcand and/or @jrafanie might have some opinions on this)
* I am not using any of the `User` helpers for Reasons™.  I can go in more depth if requested, but basically they cause more trouble (and code) then it is worth.
* Added unit tests that confirm the functionality of the attribute (in the second commit), so that could use a look.
* As noted there and in a commit message, the functionality of `owned_by_current_ldap_group filter` is somewhat counter intuitive.  When requesting it with "true" (meaning "I want the objects that are part of my ldap group"), it works as expected, but when passing false, it will only return objects that have a group, not "all other objects that aren't a part of your LDAP group".  This seems a bit weird, but I didn't fix it because the original logic that confirms this has been there a while, and changing both the ruby method and the virt attribute seemed like the wrong choice (we can update this later if we feel it is the wrong logic).
* Updated the rest of the controllers that I was able to find that would need updated to support this change (that I was able to confirm used this).  This was rebased into the third commit.


### EXTRA INFO (don't read if you don't care)

Expanding on the current logic of `owned_by_current_ldap_group`, this is what would need to change to support the logic **I think** is intuitive:

```diff
diff --git a/app/models/mixins/ownership_mixin.rb b/app/models/mixins/ownership_mixin.rb
index 90708fa..abc38da 100644
--- a/app/models/mixins/ownership_mixin.rb
+++ b/app/models/mixins/ownership_mixin.rb
@@ -13,23 +13,41 @@ module OwnershipMixin
     virtual_column :owned_by_current_ldap_group,          :type => :boolean,    :uses => :owning_ldap_group
 
     # Determine whether to return vms owned by the current user or not
-    # Requires `with_miq_group` scope to function
+    # Requires `with_miq_group` scope to function, and the COALESCE is there
+    # incase a miq_group doesn't exist for the model (since we are doing a left
+    # join in `with_miq_group`)
     #
-    # Resulting SQL: "(LOWER(miq_groups.description) = 'some_miq_group'"
+    # Resulting SQL: "(COALESCE(LOWER("miq_groups"."description"), '')) = 'some_miq_group')"
     #
     # Will result in the following when used with MiqExpression:
     #
-    #   WHERE ((LOWER("miq_groups"."description") = 'some_miq_group') = 'true')
+    #   WHERE ((COALESCE(LOWER("miq_groups"."description"), '') = 'some_miq_group') = 'true')
+    #
     virtual_attribute :owned_by_current_ldap_group, :boolean, :arel => (lambda do |t|
       ldap_group = User.current_user.try(:ldap_group).to_s.downcase
 
-      t.grouping(MiqGroup.arel_table[:description].lower.eq(ldap_group))
+      empty_string = Arel::Nodes::SqlLiteral.new "''"
+      description  = MiqGroup.arel_table[:description].lower
+      coalesce     = Arel::Nodes::NamedFunction.new "COALESCE", [description, empty_string]
+      t.grouping(coalesce.eq(ldap_group))
     end)
   end
 
   module ClassMethods
+
+    # Does a left join on the miq_groups table, but only selects from the
+    # current table.  Generated sql should look something like this:
+    #
+    #   SELECT vms.*
+    #   FROM [ownership_table]
+    #   LEFT OUTER JOIN miq_groups ON [ownership_table].miq_group_id = miq_group.id
+    #
     def with_miq_group
-      select(arel_table[Arel.star]).joins(:miq_group)
+      miq_group_table  = MiqGroup.arel_table
+      group_left_join  = arel_table.join(miq_group_table, Arel::Nodes::OuterJoin)
+                           .on(arel_table[:miq_group_id].eq(miq_group_table[:id]))
+                           .join_sources
+      select(arel_table[Arel.star]).joins(group_left_join)
     end
 
     def set_ownership(ids, options)
@@ -92,7 +110,6 @@ module OwnershipMixin
   end
 
   def owned_by_current_ldap_group
-    ldap_group = User.current_user.try(:ldap_group)
-    ldap_group && owning_ldap_group && (owning_ldap_group.downcase == ldap_group.downcase)
+    owning_ldap_group.to_s.downcase == ldap_group.to_s.downcase
   end
 end
```

The above isn't fool proof, so I don't want to add it, but I thought I would bring it to people's attention incase it was actually unintended behavior (and to justify (slightly) why I was taking so long with the tests...).