@NickLaMuro 2016-08-15
^ Go home `rubocop`, you are drunk

_These are all a `WONTFIX` since I didn't do much more than add a single line to this method.  We might want to tweak how the bot handles this, since I wasn't getting these errors when using `rubocop-git` against master (different than what we do, I realize, but it is the closest/simplest way to replicate MiqBot locally), but that is something to discuss somewhere else._