@NickLaMuro 2016-08-15
There has been even some more churn on this PR since https://github.com/ManageIQ/manageiq/pull/10290#issuecomment-239031174 ... sigh...

* Moved the `ownership_mixin` tests to a `shared_example` so they can be reused in the models that use them
  - This means these tests are now repeated 3 times for each model
  - Added an alias in the spec helper to make the `include` for the shared_example read a little better
  - Removes duplication, but definitely makes it a bit harder to grok
* `set_named_scope` is now MUCH more complex.  It is basically needs to evaluate any `MiqExpression` that might be set, and a MiqExpression can have multiple fields it is searching on (unfortunately).
  - This makes use of `MiqExpression::get_cols_from_expression` in `miq_search_exp_fields`, which now returns an array of fields.
  - This is unfortunately a bit slower than previous iterations, but not really noticeable since this method is only run once.
  - There have been a bunch of iterations on these controller methods, that are not shown because of rebases...
* `set_named_scope` is added as a `before_action` only for the explorer route.  Because of how `tree_select` sets the `@edit` and `session[:edit]` within the action, this was require to add this line in the method instead of as a `before_action`
* `set_named_scope` has a `case` statement now.
  - This is first to limit setting that scope to only when we need it for the MiqExpression, not all the time
  - This also sets up for a future PR where I will need to use a different scope for a different filter.
  - The `proc` + `includes?` is kinda smelly, but I couldn't think of a more terse way of representing that.  Suggestions welcome.